1_docker_sql/requirements.txt
pandas
google-cloud-bigquery
scikit-learn
joblib
db-dtypes
apache-airflow[gcp]
